pyproject.toml
[tool.poetry]
name = "agl-home-django"
version = "0.1.0"
description = ""
authors = ["Thomas J. Lux"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python =  ">3.10,<3.13"
# Django = "^5.0.1"
# opencv-python = "~4"
# ffmpeg-python = "^0.2.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
django-bootstrap5 = "^23.4"
agl-base-db = "^0.1.53"
opencv-python = "^4.9.0.80"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
poetry = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-server = "agl_home_django.runserver:start_django"
